Texture: split texture.cpp into own module todo
