+++ inexor/engine/octarender.cpp
@@ -1,6 +1,7 @@
 // octarender.cpp: fill vertex arrays with different cube surfaces.
 
 #include "inexor/engine/engine.hpp"
+#include "inexor/texture/cubemap.hpp"
 
 struct vboinfo
 {
