+++ inexor/engine/rendersky.cpp
@@ -1,5 +1,6 @@
 #include "inexor/engine/engine.hpp"
 #include "inexor/shared/filesystem.hpp"
+#include "inexor/texture/cubemap.hpp"
 
 Texture *sky[6] = { 0, 0, 0, 0, 0, 0 }, *clouds[6] = { 0, 0, 0, 0, 0, 0 };
