+++ inexor/texture/additionaltools.hpp
@@ -1,0 +1,7 @@
+
+#ifndef _TEX_ADDITIONALTOOLS_H
+#define _TEX_ADDITIONALTOOLS_H
+
+
+
+#endif //_TEX_ADDITIONALTOOLS_H
