+++ inexor/texture/compressedtex.hpp
@@ -1,0 +1,71 @@
+/// @file compressed texture (mostly dds) loading.
+
+#ifndef INEXOR_TEX_COMPRESSED_H
+#define INEXOR_TEX_COMPRESSED_H
+
+#include "inexor/engine/engine.hpp"
+
+extern bool loaddds(const char *filename, ImageData &image, int force = 0);
+
+enum
+{
+    DDSD_CAPS = 0x00000001,
+    DDSD_HEIGHT = 0x00000002,
+    DDSD_WIDTH = 0x00000004,
+    DDSD_PITCH = 0x00000008,
+    DDSD_PIXELFORMAT = 0x00001000,
+    DDSD_MIPMAPCOUNT = 0x00020000,
+    DDSD_LINEARSIZE = 0x00080000,
+    DDSD_BACKBUFFERCOUNT = 0x00800000,
+    DDPF_ALPHAPIXELS = 0x00000001,
+    DDPF_FOURCC = 0x00000004,
+    DDPF_INDEXED = 0x00000020,
+    DDPF_ALPHA = 0x00000002,
+    DDPF_RGB = 0x00000040,
+    DDPF_COMPRESSED = 0x00000080,
+    DDPF_LUMINANCE = 0x00020000,
+    DDSCAPS_COMPLEX = 0x00000008,
+    DDSCAPS_TEXTURE = 0x00001000,
+    DDSCAPS_MIPMAP = 0x00400000,
+    DDSCAPS2_CUBEMAP = 0x00000200,
+    DDSCAPS2_CUBEMAP_POSITIVEX = 0x00000400,
+    DDSCAPS2_CUBEMAP_NEGATIVEX = 0x00000800,
+    DDSCAPS2_CUBEMAP_POSITIVEY = 0x00001000,
+    DDSCAPS2_CUBEMAP_NEGATIVEY = 0x00002000,
+    DDSCAPS2_CUBEMAP_POSITIVEZ = 0x00004000,
+    DDSCAPS2_CUBEMAP_NEGATIVEZ = 0x00008000,
+    DDSCAPS2_VOLUME = 0x00200000,
+    FOURCC_DXT1 = 0x31545844,
+    FOURCC_DXT2 = 0x32545844,
+    FOURCC_DXT3 = 0x33545844,
+    FOURCC_DXT4 = 0x34545844,
+    FOURCC_DXT5 = 0x35545844
+};
+
+struct DDCOLORKEY { uint dwColorSpaceLowValue, dwColorSpaceHighValue; };
+struct DDPIXELFORMAT
+{
+    uint dwSize, dwFlags, dwFourCC;
+    union { uint dwRGBBitCount, dwYUVBitCount, dwZBufferBitDepth, dwAlphaBitDepth, dwLuminanceBitCount, dwBumpBitCount, dwPrivateFormatBitCount; };
+    union { uint dwRBitMask, dwYBitMask, dwStencilBitDepth, dwLuminanceBitMask, dwBumpDuBitMask, dwOperations; };
+    union { uint dwGBitMask, dwUBitMask, dwZBitMask, dwBumpDvBitMask; struct { ushort wFlipMSTypes, wBltMSTypes; } MultiSampleCaps; };
+    union { uint dwBBitMask, dwVBitMask, dwStencilBitMask, dwBumpLuminanceBitMask; };
+    union { uint dwRGBAlphaBitMask, dwYUVAlphaBitMask, dwLuminanceAlphaBitMask, dwRGBZBitMask, dwYUVZBitMask; };
+
+};
+struct DDSCAPS2 { uint dwCaps, dwCaps2, dwCaps3, dwCaps4; };
+struct DDSURFACEDESC2
+{
+    uint dwSize, dwFlags, dwHeight, dwWidth;
+    union { int lPitch; uint dwLinearSize; };
+    uint dwBackBufferCount;
+    union { uint dwMipMapCount, dwRefreshRate, dwSrcVBHandle; };
+    uint dwAlphaBitDepth, dwReserved, lpSurface;
+    union { DDCOLORKEY ddckCKDestOverlay; uint dwEmptyFaceColor; };
+    DDCOLORKEY ddckCKDestBlt, ddckCKSrcOverlay, ddckCKSrcBlt;
+    union { DDPIXELFORMAT ddpfPixelFormat; uint dwFVF; };
+    DDSCAPS2 ddsCaps;
+    uint dwTextureStage;
+};
+
+#endif //INEXOR_TEX_COMPRESSED_H
