+++ inexor/texture/texsettings.cpp
@@ -1,0 +1,20 @@
+/// @file Settings for texture loading and handling.
+
+#include "inexor/engine/engine.hpp"
+#include "inexor/texture/texsettings.hpp"
+
+VAR(hwtexsize, 1, 0, 0);
+VAR(hwcubetexsize, 1, 0, 0);
+VAR(hwmaxaniso, 1, 0, 0);
+VARFP(maxtexsize, 0, 0, 1 << 12, initwarning("texture quality", INIT_LOAD));
+VARFP(reducefilter, 0, 1, 1, initwarning("texture quality", INIT_LOAD));
+VARFP(texreduce, 0, 0, 12, initwarning("texture quality", INIT_LOAD));
+VARFP(trilinear, 0, 1, 1, initwarning("texture filtering", INIT_LOAD));
+VARFP(bilinear, 0, 1, 1, initwarning("texture filtering", INIT_LOAD));
+VARFP(aniso, 0, 0, 16, initwarning("texture filtering", INIT_LOAD));
+
+
+VARFP(texcompress, 0, 1 << 10, 1 << 12, initwarning("texture quality", INIT_LOAD));
+VARFP(texcompressquality, -1, -1, 1, setuptexcompress());
+
+VARFP(usenp2, 0, 0, 1, initwarning("texture quality", INIT_LOAD));
